controller modifs- admin

+++ backend/controllers/user.js
@@ -91,11 +91,16 @@ exports.adminDeleteUser = ( req , res ) => {
     if (!user) {
       return res.status(404).json({ error: "Utilisateur non trouvée !" });
     }
+    if(user.role === "admin"){
     User.deleteOne({ _id: req.params.id })
       .then((user) =>
         res.status(200).json({ message: "Utilisateur supprimée par administrateur !" })
       )
       .catch((error) => res.status(403).json({ error }));
+    }
+      else{
+        return res.status(403).json({error:"Vous n'êtes pas administrateur "});
+      }
   })
   .catch((error) => res.status(400).json({ error }));
 }
